+++ frontend/src/pages/Login/index.tsx
@@ -1,8 +1,8 @@
 import Card from '@/components/@common/Card/Card';
 import Loading from '@/components/@common/Loading/Loading';
+import LoginButton from '@/components/login/LoginButton/LoginButton';
 import { mobileTitleSecondary } from '@/constants/titleType';
 import useGetLoginURL from '@/hooks/login/useGetLoginURL';
-import LoginButton from '@/pages/Login/LoginButton/LoginButton';
 import * as S from '@/pages/Login/index.styles';
 import { LoginCardStyle } from '@/styles/cardStyle';
