+++ frontend/src/storybook/Login/LoginButton.stories.tsx
@@ -1,6 +1,6 @@
 import { PropsWithStrictChildren } from 'gongseek-types';
 
-import LoginButton, { LoginButtonProps } from '@/pages/Login/LoginButton/LoginButton';
+import LoginButton, { LoginButtonProps } from '@/components/login/LoginButton/LoginButton';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story } from '@storybook/react';
